Add cast. Fix one incorrect reference.

+++ src/entropy/ExpGolombDecoder.cpp
@@ -26,7 +26,7 @@ ExpGolombDecoder::ExpGolombDecoder(InputBitStream& bitstream, bool sgn)
 int ExpGolombDecoder::decode(byte block[], uint blkptr, uint len)
 {
     byte* buf = &block[blkptr];
-    const uint len8 = len & -8;
+    const uint len8 = len & uint(-8);
 
     for (uint i = 0; i < len8; i+=8) {
         buf[i]   = decodeByte();
